refactor: split native command modules
