+++ packages/stim-cli/src/commands/dev-client.ts
@@ -1,0 +1,81 @@
+import { join } from 'node:path';
+import { readFileSync } from 'node:fs';
+import { getExecutor } from '../exec.ts';
+import { isPackageResolvable } from '../project.ts';
+
+export function devClientScheme(
+  root: string,
+  appPath: string | null = null,
+  { exec = null }: { exec?: import('../exec.ts').Executor | null } = {},
+): string | undefined {
+  if (!hasDevClient(root)) return undefined;
+  const fromBundle = pickDevClientScheme(readBundleSchemes(appPath, { exec }));
+  if (fromBundle) return fromBundle;
+  const app = readJson(join(root, 'app.json')) as { expo?: { scheme?: unknown }; scheme?: unknown } | null;
+  const raw = app?.expo?.scheme ?? app?.scheme ?? null;
+  const scheme = Array.isArray(raw) ? raw.find((s) => typeof s === 'string' && s.trim() !== '') : raw;
+  if (typeof scheme !== 'string' || scheme.trim() === '') return undefined;
+  return scheme.trim();
+}
+
+export function schemesFromInfoPlist(plist: unknown): string[] {
+  const types = (plist as { CFBundleURLTypes?: unknown })?.CFBundleURLTypes;
+  if (!Array.isArray(types)) return [];
+  const out: string[] = [];
+  for (const type of types) {
+    const schemes = (type as { CFBundleURLSchemes?: unknown })?.CFBundleURLSchemes;
+    if (Array.isArray(schemes)) out.push(...schemes.filter((s) => typeof s === 'string' && s.trim() !== ''));
+  }
+  return out;
+}
+
+export function readBundleSchemes(
+  appPath: unknown,
+  { exec = null }: { exec?: import('../exec.ts').Executor | null } = {},
+): string[] {
+  if (typeof appPath !== 'string' || appPath.trim() === '') return [];
+  const e = exec || getExecutor();
+  let out;
+  try {
+    out = e.runFile('plutil', ['-convert', 'json', '-o', '-', join(appPath, 'Info.plist')]);
+  } catch {
+    return [];
+  }
+  try {
+    return schemesFromInfoPlist(JSON.parse(String(out)));
+  } catch {
+    return [];
+  }
+}
+
+const THIRD_PARTY_SCHEME =
+  /^(?:fb\d+|com\.googleusercontent\.apps\.|msauth\.|msauthv2|twitterkit-|db-[a-z0-9]+$|spotify|snapchat|com\.facebook)/i;
+
+export function pickDevClientScheme(schemes: unknown): string | null {
+  const all = (Array.isArray(schemes) ? schemes : [])
+    .filter((s) => typeof s === 'string' && s.trim() !== '')
+    .map((s) => s.trim())
+    .filter((s) => !/^(?:https?|mailto|tel|sms|itms(?:-apps)?)$/i.test(s));
+  const expo = all.filter((s) => s.startsWith('exp+'));
+  const pool = expo.length ? expo : all.filter((s) => !THIRD_PARTY_SCHEME.test(s));
+  const sorted = pool.toSorted((a, b) => b.length - a.length);
+  return sorted[0] ?? null;
+}
+
+function hasDevClient(root: string): boolean {
+  const pkg = readJson(join(root, 'package.json')) as {
+    dependencies?: Record<string, unknown>;
+    devDependencies?: Record<string, unknown>;
+  } | null;
+  const deps = { ...pkg?.dependencies, ...pkg?.devDependencies };
+  if ('expo-dev-client' in deps) return true;
+  return isPackageResolvable(root, 'expo-dev-client');
+}
+
+function readJson(file: string): unknown {
+  try {
+    return JSON.parse(readFileSync(file, 'utf-8'));
+  } catch {
+    return null;
+  }
+}
